+++ ChangeLog.md
@@ -4,4 +4,5 @@
 
 ### Version 1.0
 
+- (b.2) Parallel integral calculation
 - (b.1) Sequential integral calculation